Use kerboscript stack when calling kerboscript functions from lua
Reverted some changes from a104d58b867e46a92e8fda43baeac405acee8386 commit that created a custom stack operator for lua
Now kerboscript and lua use the same stack when calling kerboscript functions and SuffixDelegateResults
