+++ src/kOS.Safe/Function/SafeFunctionBase.cs
@@ -8,7 +8,6 @@ namespace kOS.Safe.Function
 {
     public abstract class SafeFunctionBase
     {
-        public IStackOperator stackOperator;
         public string functionName;
 
         /// <summary>
@@ -104,7 +103,14 @@ namespace kOS.Safe.Function
         /// If the assert fails, an exception is thrown.
         /// </summary>
         /// <param name="shared"></param>
-        protected void AssertArgBottomAndConsume(SafeSharedObjects shared) => stackOperator.AssertArgBottomAndConsume(shared, functionName);
+        protected void AssertArgBottomAndConsume(SafeSharedObjects shared)
+        {
+            object shouldBeBottom = shared.Cpu.PopArgumentStack();
+            if (shouldBeBottom != null && shouldBeBottom.GetType() == OpcodeCall.ArgMarkerType)
+                return; // Assert passed.
+
+            throw new KOSArgumentMismatchException("Too many arguments were passed to " + functionName);
+        }
 
         /// <summary>
         /// A utility function that a function's Execute() may use if it wishes to, to get a count of
@@ -112,7 +118,24 @@ namespace kOS.Safe.Function
         /// </summary>
         /// <param name="shared"></param>
         /// <returns>Number of args as yet unpopped.  returns zero if there are no args, or -1 if there's a bug and the argstart marker is missing.</returns>
-        protected int CountRemainingArgs(SafeSharedObjects shared) => stackOperator.CountRemainingArgs(shared);
+        protected int CountRemainingArgs(SafeSharedObjects shared)
+        {
+            int depth = 0;
+            bool found = false;
+            bool stillInStack = true;
+            while (stillInStack && !found)
+            {
+                object peekItem = shared.Cpu.PeekRawArgument(depth, out stillInStack);
+                if (stillInStack && peekItem != null && peekItem.GetType() == OpcodeCall.ArgMarkerType)
+                    found = true;
+                else
+                    ++depth;
+            }
+            if (found)
+                return depth;
+            else
+                return -1;
+        }
 
         /// <summary>
         /// A utility function that a function's Execute() should use in place of cpu.PopValue(),
@@ -120,7 +143,13 @@ namespace kOS.Safe.Function
         /// is, it will throw the appropriate error.
         /// </summary>
         /// <returns></returns>
-        protected object PopValueAssert(SafeSharedObjects shared, bool barewordOkay = false) => stackOperator.PopValueAssert(shared, functionName, barewordOkay);
+        protected object PopValueAssert(SafeSharedObjects shared, bool barewordOkay = false)
+        {
+            object returnValue = shared.Cpu.PopValueArgument(barewordOkay);
+            if (returnValue != null && returnValue.GetType() == OpcodeCall.ArgMarkerType)
+                throw new KOSArgumentMismatchException("Too few arguments were passed to " + functionName);
+            return returnValue;
+        }
 
         /// <summary>
         /// A utility function that a function's Execute() should use in place of cpu.PopArgumentStack(),
@@ -128,7 +157,13 @@ namespace kOS.Safe.Function
         /// is, it will throw the appropriate error.
         /// </summary>
         /// <returns></returns>
-        protected object PopStackAssert(SafeSharedObjects shared) => stackOperator.PopStackAssert(shared, functionName);
+        protected object PopStackAssert(SafeSharedObjects shared)
+        {
+            object returnValue = shared.Cpu.PopArgumentStack();
+            if (returnValue != null && returnValue.GetType() == OpcodeCall.ArgMarkerType)
+                throw new KOSArgumentMismatchException("Too few arguments were passed to " + functionName);
+            return returnValue;
+        }
 
         /// <summary>
         /// Identical to PopValueAssert, but with the additional step of coercing the result
