+++ src/kOS/Lua/Binding.cs
@@ -337,138 +337,26 @@ namespace kOS.Lua
             var state = KeraLua.Lua.FromIntPtr(L);
             var binding = bindings[state.MainThread.Handle];
             var ksFunction = binding.objects[state.ToUserData(1)];
+            
+            var stack = (binding.Shared.Cpu as LuaCPU).Stack;
+            stack.Clear();
+            stack.PushArgument(new KOSArgMarkerType());
+            for (int i = 2; i <= state.GetTop(); i++)
+                stack.PushArgument(ToCSharpObject(state, i, binding));
+            
             if (ksFunction is SafeFunctionBase function)
             {
-                var stackOperator = function.stackOperator as LuaStackOperator;
-                stackOperator.stack.Clear();
-                for (int i = 2; i <= state.GetTop(); i++)
-                    stackOperator.stack.Push(ToCSharpObject(state, i, binding));
-
                 try { function.Execute(binding.Shared); }
                 catch (Exception e) { Debug.Log(e); return state.Error(e.Message); }
-
                 return PushLuaType(state, Structure.ToPrimitive(function.ReturnValue), binding);
             }
-            var delegateResult = ksFunction as DelegateSuffixResult;
-            if (delegateResult == null)
-                return state.Error(string.Format("attempt to call a non function {0} value", ksFunction.GetType().Name));
-
-            // the next entire bit is copied from DelegateSuffixResult.Invoke method with a few changes
-            // that make it access arguments from the lua stack instead of kerboscript stack
-            try
+            if (ksFunction is DelegateSuffixResult delegateResult)
             {
-                var delInfo = delegateResult.RawDelInfo;
-                var args = new List<object>();
-                var paramArrayArgs = new List<Structure>();
-
-                // Will be true iff the lastmost parameter of the delegate is using the C# 'param' keyword and thus
-                // expects the remainder of the arguments marshalled together into one array object.
-                bool isParamArrayArg = false;
-
-                //CpuUtility.ReverseStackArgs(cpu, false);
-                int popIndex = 2;
-                for (int i = 0; i < delInfo.Parameters.Length; ++i)
-                {
-                    DelegateParameter paramInfo = delInfo.Parameters[i];
-
-                    //object arg = cpu.PopValueArgument();
-                    object arg = popIndex <= state.GetTop()? ToCSharpObject(state, popIndex++, binding) : new KOSArgMarkerType();
-                    Type argType = arg.GetType();
-                    isParamArrayArg = i == delInfo.Parameters.Length - 1 && delInfo.Parameters[i].IsParams;
-
-                    if (arg != null && arg.GetType() == CpuUtility.ArgMarkerType)
-                    {
-                        if (isParamArrayArg)
-                            break; // with param arguments, you want to consume everything to the arg bottom - it's normal.
-                        else
-                            throw new KOSArgumentMismatchException(delInfo.Parameters.Length, delInfo.Parameters.Length - (i + 1));
-                    }
-
-                    // Either the expected type of this one parameter, or if it's a 'param' array as the last arg, then
-                    // the expected type of that array's elements:
-                    Type paramType = (paramInfo.IsParams ? paramInfo.ParameterType.GetElementType() : paramInfo.ParameterType);
-
-                    // Parameter type-safe checking:
-                    bool inheritable = paramType.IsAssignableFrom(argType);
-                    if (!inheritable)
-                    {
-                        bool castError = false;
-                        // If it's not directly assignable to the expected type, maybe it's "castable" to it:
-                        try
-                        {
-                            arg = Convert.ChangeType(arg, Type.GetTypeCode(paramType));
-                        }
-                        catch (InvalidCastException)
-                        {
-                            throw new KOSCastException(argType, paramType);
-                        }
-                        catch (FormatException)
-                        {
-                            castError = true;
-                        }
-                        if (castError)
-                        {
-                            throw new Exception(string.Format("Argument {0}({1}) to method {2} should be {3} instead of {4}.", (delInfo.Parameters.Length - i), arg, delInfo.Name, paramType.Name, argType));
-                        }
-                    }
-
-                    if (isParamArrayArg)
-                    {
-                        paramArrayArgs.Add(Structure.FromPrimitiveWithAssert(arg));
-                        --i; // keep hitting the last item in the param list again and again until a forced break because of arg bottom marker.
-                    }
-                    else
-                    {
-                        args.Add(Structure.FromPrimitiveWithAssert(arg));
-                    }
-                }
-                if (isParamArrayArg)
-                {
-                    // collect the param array args that were at the end into the one single
-                    // array item that will be sent to the method when invoked:
-                    args.Add(paramArrayArgs.ToArray());
-                }
-                // Consume the bottom marker under the args, which had better be
-                // immediately under the args we just popped, or the count was off.
-                if (!isParamArrayArg) // A param array arg will have already consumed the arg bottom mark.
-                {
-                    bool foundArgMarker = false;
-                    int numExtraArgs = 0;
-                    //while (cpu.GetArgumentStackSize() > 0 && !foundArgMarker)
-                    while (state.GetTop()-popIndex >= 0 && !foundArgMarker)
-                    {
-                        //object marker = cpu.PopValueArgument();
-                        object marker = popIndex <= state.GetTop()? ToCSharpObject(state, popIndex++, binding) : new KOSArgMarkerType();
-                        if (marker != null && marker.GetType() == CpuUtility.ArgMarkerType)
-                            foundArgMarker = true;
-                        else
-                            ++numExtraArgs;
-                    }
-                    if (numExtraArgs > 0)
-                        throw new KOSArgumentMismatchException(delInfo.Parameters.Length, delInfo.Parameters.Length + numExtraArgs);
-                }
-
-                // Delegate.DynamicInvoke expects a null, rather than an array of zero length, when
-                // there are no arguments to pass:
-                object[] argArray = (args.Count > 0) ? args.ToArray() : null;
-
-                //object val = call(argArray);
-                object val = delegateResult.RawCall(argArray);
-                if (delInfo.ReturnType == typeof(void))
-                {
-                    //value = ScalarValue.Create(0);
-                    delegateResult.RawSetValue(ScalarValue.Create(0));
-                }
-                else
-                {
-                    delegateResult.RawSetValue(Structure.FromPrimitiveWithAssert(val));
-                }
+                try { delegateResult.Invoke(binding.Shared.Cpu); }
+                catch (Exception e) { Debug.Log(e); return state.Error(e.Message); }
                 return PushLuaType(state, Structure.ToPrimitive(delegateResult.Value), binding);
-            } catch (Exception e)
-            {
-                Debug.Log(e);
-                return state.Error(e.Message);
             }
+            return state.Error(string.Format("attempt to call a non function {0} value", ksFunction.GetType().Name));
         }
 
         private static int PushLuaType(KeraLua.Lua state, object obj, BindingData binding)
